changed type of lat / long fields to double

+++ VideoServer/src/org/linesofcode/videoServer/Broadcast.java
@@ -7,22 +7,22 @@ public class Broadcast implements Serializable {
 	private static final long serialVersionUID = 5921215050421366377L;
 	
 	private String id;
-	private long lattitude;
-	private long longitude;
+	private double lattitude;
+	private double longitude;
 	private String title;
 	
-	public Broadcast(String id, long lattitude, long longitude, String title) {
+	public Broadcast(String id, double lattitude, double longitude, String title) {
 		this.id = id;
 		this.lattitude = lattitude;
 		this.longitude = longitude;
 		this.title = title;
 	}
 
-	public void setLattitude(long lattitude) {
+	public void setLattitude(double lattitude) {
 		this.lattitude = lattitude;
 	}
 
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 
@@ -30,11 +30,11 @@ public class Broadcast implements Serializable {
 		return id;
 	}
 
-	public long getLattitude() {
+	public double getLattitude() {
 		return lattitude;
 	}
 
-	public long getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
